Extract the form to it's own file
